Fix problem with new device hierarchy discovery
The device import function always takes the first sub-class of CanProcess it finds that is not in core.processor. But with the new device base classes it also needs to be made sure that Device and StateMachineDevice are not used, they are in a different module (devices). That probably went unnoticed because the other devices were in an alphabetically advantageous order.