Simulation can construct its own ControlServer
Simulation now accepts different input, most importantly a host:port string
from which a ControlServer is constructed. For consistency this is also
available via the property setter.

For better control over when the server starts listening, it has to be
started explicitly now, either via calling start_server or via a new
parameter of the constructor called start.

A few more tests have been added around this. Some of the exceptions may
be seen by users, those will be addressed as part of #135 (or related).

Fixes #96.