Add release notes to generated documentation
This adds release notes to the sphinx generated documentation so
that it becomes part of the source and the online docs.

The current release has the :orphan: marker at the top, because it is
not included in the TOC yet, because it's work in progress and not
yet intended to be seen by users.